+++ packages/node/src/indexer/fetch.service.ts
@@ -152,13 +152,20 @@ function callFilterToQueryEntry(
     );
   }
 
-  if (filter.function) {
+  if (filter.function === null || filter.function === '0x') {
+    conditions.push({
+      field: 'func',
+      value: true,
+      matcher: 'isNull',
+    });
+  } else if (filter.function) {
     conditions.push({
       field: 'func',
       value: functionToSighash(filter.function),
       matcher: 'equalTo',
     });
   }
+
   return {
     entity: 'evmTransactions',
     conditions,
@@ -186,7 +193,7 @@ export function buildDictionaryQueryEntries(
           if (
             filter.from !== undefined ||
             filter.to !== undefined ||
-            filter.function
+            filter.function !== undefined
           ) {
             queryEntries.push(callFilterToQueryEntry(filter, ds.options));
           } else {